+++ data_models/ads_decision_criteria_data_models.py
@@ -42,7 +42,7 @@ else:
 #   note: - _pk and audit columns come from BaseInfoMixin
 # ============================================
 class ads_decision_criteria(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_decision_criteria'
 
     sales_project_fk = sa.Column(sa.String(36), sa.ForeignKey('salesproject._pk', ondelete='CASCADE', onupdate='CASCADE'), nullable=False, unique=True, index=True)
@@ -102,7 +102,7 @@ class ads_decision_criteria(BaseModel, BaseInfoMixin):
 #   note: - _pk and audit columns come from BaseInfoMixin
 # ============================================
 class ads_decision_criteria_details_1criterialist(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_decision_criteria_details_1criterialist'
     ads_decision_criteria_fk = sa.Column(sa.String(36), sa.ForeignKey('ads_decision_criteria._pk', ondelete='CASCADE', onupdate='CASCADE'), nullable=False, index=True)
     ads_decision_criteria_data = relationship("ads_decision_criteria", back_populates="ads_decision_criteria_details_1criterialist_data") # to TOP
@@ -154,7 +154,7 @@ class ads_decision_criteria_details_1criterialist(BaseModel, BaseInfoMixin):
 #   note: - _pk and audit columns come from BaseInfoMixin
 # ============================================
 class ads_decision_criteria_details_2maptoperson(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_decision_criteria_details_2maptoperson'
 
     # to TOP: DECISION CRITERIA OBJECT (ads_decision_criteria_details_1criterialist)