+++ data_models/ads_evaluation_data_models.py
@@ -42,7 +42,7 @@ else:
 
 
 class ads_evaluation(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_evaluation'
 
     sales_project_fk = sa.Column(sa.String(36), sa.ForeignKey(