+++ data_models/ads_relationships_data_models.py
@@ -41,7 +41,7 @@ else:
 #   note: - _pk and audit columns come from BaseInfoMixin
 # ============================================
 class ads_relationships(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_relationships'
 
     sales_project_fk = sa.Column(sa.String(36), sa.ForeignKey('salesproject._pk', ondelete='CASCADE', onupdate='CASCADE'), nullable=False, unique=True, index=True)
@@ -101,7 +101,7 @@ class ads_relationships(BaseModel, BaseInfoMixin):
 #   note: - _pk and audit columns come from BaseInfoMixin
 # ============================================
 class ads_relationships_details(BaseModel, BaseInfoMixin):
-    __versioned__ = {} # activate data hostory and versioning mechanism of SQLAlchemy Continuum
+    __versioned__ = {} # activate data history and versioning mechanism of SQLAlchemy Continuum
     __tablename__ = 'ads_relationships_details'
 
     ads_relationships_fk = sa.Column(sa.String(36), sa.ForeignKey('ads_relationships._pk', ondelete='CASCADE', onupdate='CASCADE'), nullable=False, index=True)